Moved the SignalManager to managers.py
